portfolio admin form functions repaired

+++ portfolio_manage.php
@@ -6,7 +6,9 @@ require 'functions.php';
 $db = new PDO($dsn, $user);
 $db->setAttribute(PDO::ATTR_DEFAULT_FETCH_MODE, PDO::FETCH_ASSOC);
 
-deleteProject($db, $_POST);
+if ($_POST['delete'] != NULL && $_POST['id'] != NULL) {
+    deleteProject($db, $_POST);
+}
 
 $singleProject = getSingleProject($db, $_POST);
 
@@ -26,15 +28,16 @@ $singleProject = getSingleProject($db, $_POST);
         <h2>EDIT PROJECT</h2>
         <form method="post" action="portfolio_update.php">
             Project Description:<br>
-            <input type="text" name="projectDescription" value="<?php echo $singleProject['projectDescription']?>"><br>
+            <input type="text" name="projectDescription" value="<?php echo $singleProject['projectDescription']; ?>"><br>
             Link:<br>
-            <input type="text" name="link" value="<?php echo $singleProject['link']?>"><br>
+            <input type="text" name="link" value="<?php echo $singleProject['link']; ?>"><br>
             Image Source:<br>
-            <input type="text" name="imageSource" value="<?php echo $singleProject['imageSource']?>"><br>
+            <input type="text" name="imageSource" value="<?php echo $singleProject['imageSource']; ?>"><br>
             Alternative Image Text:<br>
-            <input type="text" name="alternativeText" value="<?php echo $singleProject['alternativeText']?>"><br><br>
+            <input type="text" name="alternativeText" value="<?php echo $singleProject['alternativeText']; ?>"><br><br>
             <input type="submit" value="Save">
-            <input type="hidden" name="id" value="<?php echo $_POST['id']?>">
+            <input type='hidden' name='edit' value='Edit'>
+            <input type="hidden" name="id" value="<?php echo $singleProject['id']; ?>">
         </form>
     </div>
     <div class="container">
